requirements.txt
# Core dependencies
textual==0.88.1
pygame-ce==2.5.2

# Utility libraries
tinytag==2.0.0

# Testing
pytest==8.3.4
